+++ docs/changes.rst
@@ -20,6 +20,11 @@ Pending
   "Django Debug Toolbar was unable to parse value." when there's a decoding
   error.
 * Updated docs to show incompatibility with Django Channels.
+* Hide the toolbar's migrations unless ``debug_toolbar.store.DatabaseStore``
+  is being used. This may change in the future.
+* Hide ``debug_toolbar.HistoryEntry`` as a model unless
+  ``debug_toolbar.store.DatabaseStore`` is being used. This may change in the
+  future.
 
 6.0.0 (2025-07-22)
 ------------------